Complete App Rewrite

+++ Sudachi/Sudachi/Core/core/hle/kernel/k_page_table_base.cpp
@@ -97,7 +97,7 @@ void ClearBackingRegion(Core::System& system, KPhysicalAddress addr, u64 size, u
                                                         size, fill_value);
 }
 
-void ClearBackingRegion2(Core::System& system, KPhysicalAddress addr, u16 size, u8 fill_value) {
+void ClearBackingRegion2(Core::System& system, KPhysicalAddress addr, uint_fast8_t size, uint8_t fill_value) {
         system.DeviceMemory().buffer.ClearBackingRegion(GetInteger(addr) - Core::DramMemoryMap::Base,
                                                         size, fill_value);
 }
@@ -109,7 +109,7 @@ Result InvalidateDataCache(AddressType addr, u64 size) {
 }
 
 template <typename AddressType>
-Result StoreDataCache(AddressType addr, u8 size) {
+Result StoreDataCache(AddressType addr, u64 size) {
     R_SUCCEED();
 }
